feat: show app name
