+++ apps/membership_app1/pubspec.yaml
@@ -10,6 +10,10 @@ dependencies:
   flutter:
     sdk: flutter
   firebase_core: ^3.7.0
+  flutter_riverpod: ^2.6.1
+
+  core:
+    path: ../../packages/core
 
 dev_dependencies:
   flutter_test:
