+++ apps/membership_app2/lib/app.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:membership_app2/view/page/home_page.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.lightGreen),
+      ),
+      home: const HomePage(),
+    );
+  }
+}
